requirements.txt
transformers==4.32.0
accelerate
tiktoken
einops
scipy
transformers_stream_generator==0.0.4
peft
deepspeed
auto-gptq
optimum
sentence-transformers

google
googlesearch-python

selenium
gradio

pytest